Adds strptime fallback for Z remainder (#5)
It seems that the date format of the api has changed.
From `2017-01-30T00:00` to `2017-01-30T00:00Z`. 
Z means Zulu / GMT. But from what I can tell it is not GMT but appears still to be 
CET/CEST. I think it got in there by accident. So we keep the old date format 
in case the api rolls back the change